Adding and removing products from favourites has been implemented

+++ src/App.js
@@ -42,20 +42,19 @@ export const App = () => {
     }
   };
 
-  const onDeleteFavoriteItem = id => {
-    try {
+  const onAddToFavorite = async obj => {
+    if (favorites.find(favObj => favObj.id === obj.id)) {
       axios.delete(
-        `https://63c4354d8067b6bef6d59cf6.mockapi.io/favorites/${id}`
+        `https://63c4354d8067b6bef6d59cf6.mockapi.io/favorites/${obj.id}`
       );
-      setCartProducts(prev => prev.filter(item => item.id !== id));
-    } catch (error) {
-      console.log('error remove');
+      setFavorites(prev => prev.filter(item => item.id !== obj.id));
+    } else {
+      const { data } = await axios.post(
+        'https://63c4354d8067b6bef6d59cf6.mockapi.io/favorites',
+        obj
+      );
+      setFavorites(prev => [...prev, data]);
     }
-  };
-
-  const onAddToFavorite = obj => {
-    axios.post('https://63c4354d8067b6bef6d59cf6.mockapi.io/favorites', obj);
-    setFavorites(prev => [...prev, obj]);
   };
 
   const searchInput = event => {
@@ -82,7 +81,10 @@ export const App = () => {
         <Route
           path='/favorites'
           element={
-            <Favorites favorites={favorites} onRemove={onDeleteFavoriteItem} />
+            <Favorites
+              favorites={favorites}
+              onAddToFavorite={onAddToFavorite}
+            />
           }
         />
         <Route
